+++ DataCollector/src/main/java/com/braintreepayments/api/datacollector/MagnesInternalClient.kt
@@ -7,9 +7,11 @@ import androidx.annotation.RestrictTo
 import com.braintreepayments.api.core.Configuration
 import lib.android.paypal.com.magnessdk.Environment
 import lib.android.paypal.com.magnessdk.InvalidInputException
+import lib.android.paypal.com.magnessdk.MagnesResult
 import lib.android.paypal.com.magnessdk.MagnesSDK
 import lib.android.paypal.com.magnessdk.MagnesSettings
 import lib.android.paypal.com.magnessdk.MagnesSource
+import lib.android.paypal.com.magnessdk.MagnesSubmitStatus
 
 @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
 class MagnesInternalClient(
@@ -20,10 +22,20 @@ class MagnesInternalClient(
     internal fun getClientMetadataId(
         context: Context?,
         configuration: Configuration?,
-        request: DataCollectorInternalRequest?
-    ): String {
-        if (context == null || configuration == null || request == null) {
-            return ""
+        request: DataCollectorInternalRequest?,
+        callback: (String?, Exception?) -> Unit
+    ) {
+        if (context == null) {
+            callback(null, IllegalArgumentException("Context is null"))
+            return
+        }
+        if (configuration == null) {
+            callback(null, IllegalArgumentException("Configuration is null"))
+            return
+        }
+        if (request == null) {
+            callback(null, IllegalArgumentException("Request is null"))
+            return
         }
 
         val btEnvironment = configuration.environment
@@ -44,12 +56,20 @@ class MagnesInternalClient(
                     .setHasUserLocationConsent(request.hasUserLocationConsent)
 
             magnesSDK.setUp(magnesSettingsBuilder.build())
-            val result = magnesSDK.collectAndSubmit(
+            lateinit var result: MagnesResult
+            result = magnesSDK.collectAndSubmit(
                 context.applicationContext,
                 request.clientMetadataId,
                 request.additionalData
-            )
-            return result.paypalClientMetaDataId
+            ) { status, _ ->
+                // Callback is invoked when device data collection and submit API completes
+                when (status) {
+                    MagnesSubmitStatus.SUCCESS -> callback(result.paypalClientMetaDataId, null)
+                    MagnesSubmitStatus.ERROR -> callback(null, CallbackSubmitException.SubmitError())
+                    MagnesSubmitStatus.TIMEOUT -> callback(null, CallbackSubmitException.SubmitTimeout())
+                    else -> callback(null, CallbackSubmitException.Unknown(status.toString()))
+                }
+            }
         } catch (e: InvalidInputException) {
             // Either clientMetadataId or appGuid exceeds their character limit
             Log.e(
@@ -57,7 +77,7 @@ class MagnesInternalClient(
                 "Error fetching client metadata ID. Contact Braintree Support for assistance.",
                 e
             )
-            return ""
+            callback(null, e)
         }
     }
 }